+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - Add a `get_special_tokens_mask` method on the tokenizer that returns a [0, 1]
   mask of the underlying tokens.
 - Add autodoc of API for main tokenizers.
+- Install the python sentencepiece package by default. Add utilities for package
+  management.
 
 ## 0.12.0 (2020-01-25)